finalised the functions to read Excel data file

+++ headers/country.h
@@ -5,12 +5,12 @@ class country {
 public: 
 	std::string name;
 	std::string othercountryexports; /// \bug figure out how we want to read exchange rates.
-	float elasticityimports;
-	float elasticityexports;
-	float income;
-	float A;
-	float totalimports;
-	float totalexports;
+	double elasticityimports;
+	double elasticityexports;
+	double income;
+	double A;
+	double totalimports;
+	double totalexports;
 	std::string exportmarkets;
 
 	double GDP();
